refactor: createAssoc to createAssociation
