+++ app/src/main/java/com/github/se/assocify/navigation/NavigationGraph.kt
@@ -4,12 +4,12 @@ import androidx.navigation.NavGraphBuilder
 import com.github.se.assocify.model.database.AssociationAPI
 import com.github.se.assocify.model.database.UserAPI
 import com.github.se.assocify.ui.screens.chat.chatGraph
-import com.github.se.assocify.ui.screens.createAssoc.createAssoGraph
+import com.github.se.assocify.ui.screens.createAssociation.createAssoGraph
 import com.github.se.assocify.ui.screens.event.eventGraph
 import com.github.se.assocify.ui.screens.home.homeGraph
 import com.github.se.assocify.ui.screens.login.loginGraph
 import com.github.se.assocify.ui.screens.profile.profileGraph
-import com.github.se.assocify.ui.screens.selectAssoc.selectAssoGraph
+import com.github.se.assocify.ui.screens.selectAssociation.selectAssoGraph
 import com.github.se.assocify.ui.screens.treasury.treasuryGraph
 
 fun NavGraphBuilder.mainNavGraph(
