+++ app/src/main/java/com/github/se/assocify/ui/screens/createAssociation/CreateAssociationGraph.kt
@@ -1,4 +1,4 @@
-package com.github.se.assocify.ui.screens.selectAssoc
+package com.github.se.assocify.ui.screens.createAssociation
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
@@ -6,17 +6,14 @@ import com.github.se.assocify.model.database.AssociationAPI
 import com.github.se.assocify.model.database.UserAPI
 import com.github.se.assocify.navigation.Destination
 import com.github.se.assocify.navigation.NavigationActions
+import com.github.se.assocify.ui.screens.createAsso.CreateAssoScreen
 
-fun NavGraphBuilder.selectAssoGraph(
+fun NavGraphBuilder.createAssoGraph(
     navigationActions: NavigationActions,
     userAPI: UserAPI,
     associationAPI: AssociationAPI
 ) {
-  composable(route = Destination.SelectAsso.route) {
-    SelectAssociation(
-        navActions = navigationActions,
-        associationAPI = associationAPI,
-        userAPI = userAPI,
-    )
+  composable(route = Destination.CreateAsso.route) {
+    CreateAssoScreen(navigationActions, associationAPI, userAPI)
   }
 }
