+++ app/src/main/java/com/github/se/assocify/ui/screens/selectAssociation/SelectAssociationGraph.kt
@@ -1,4 +1,4 @@
-package com.github.se.assocify.ui.screens.createAssoc
+package com.github.se.assocify.ui.screens.selectAssociation
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
@@ -6,14 +6,17 @@ import com.github.se.assocify.model.database.AssociationAPI
 import com.github.se.assocify.model.database.UserAPI
 import com.github.se.assocify.navigation.Destination
 import com.github.se.assocify.navigation.NavigationActions
-import com.github.se.assocify.ui.screens.createAsso.CreateAssoScreen
 
-fun NavGraphBuilder.createAssoGraph(
+fun NavGraphBuilder.selectAssoGraph(
     navigationActions: NavigationActions,
     userAPI: UserAPI,
     associationAPI: AssociationAPI
 ) {
-  composable(route = Destination.CreateAsso.route) {
-    CreateAssoScreen(navigationActions, associationAPI, userAPI)
+  composable(route = Destination.SelectAsso.route) {
+    SelectAssociation(
+        navActions = navigationActions,
+        associationAPI = associationAPI,
+        userAPI = userAPI,
+    )
   }
 }
